Fix imageio warning for indivisible image size
Fixed the IMAGEIO FFMPEG_WRITER_WARNING that arrises due to the image
size/vispy canvas size not being divisible by the macro block size of 16.
The ability to set the canvas size for the Visualizer class
as an argument during initialization has been added and the default
canvas size is now divisible by the macro block size to deal with the
warnings and results in higher quality video outputs.