Implement add_axis method for Visualizer class
Implemented the add_axis method for the Visualizer class, which can be,
in its very basic current state, used to add an x or y axis to the
scene. Future work will be to add z axis through a transformation,
and provide parameters to specify range, color, width, font size, tick
spacing, etc.

Also, auto scaling and sizing based on spatial domain that the object
will traverse throughout the simulation is intended. This can also be
used for other features such as camera view range control.